+++ R/taxa_translate.R
@@ -27,12 +27,10 @@
 #' "DNI" (Do Not Include).  Default is NULL so no action is taken.  "NA"s are
 #' always removed.
 #'
-#' Optional parameter to `clean` the data of leading and trailing white
-#' space.   Default is FALSE (no action).  Not fully implemented.
-#'
-#' The optional parameter `match_caps` matches on all upper case (input and
-#' official lists).  The default  is FALSE and matches will be performed without
-#' any additional steps for case.  Not fully implemented.
+#' Optional parameter `trim_ws` is inused to invoke the function `trimws` to
+#' remove from the taxa matching field any leading and trailing white space.
+#' Default is FALSE (no action).  All horizontal and vertical white space
+#' characters are removed.  See ?trimws for additional information.
 #'
 #' The taxa list and metadata file names will be added to the results as two
 #' new columns.
@@ -60,10 +58,8 @@
 #' @param sum_n_taxa_group_by Column names for user data to use for grouping the
 #' data when summarizing the user data.  Suggestions are SAMPID and TAXA_ID.
 #' Default = NULL
-#' @param clean Should the taxa have leading and trailing white space removed.
+#' @param trim_ws Should the taxa have leading and trailing white space removed.
 #' Non-braking spaces (e.g., from ITIS) also removed. Default = FALSE
-#' @param match_caps Should the matching be performed using ALL CAPS.
-#' Default = FALSE
 #'
 #' @return A list with four elements.  The first (merge) is the user data frame
 #' with additional columns from the official data appended to it.  Names from
@@ -170,8 +166,7 @@ taxa_translate <- function(df_user = NULL
                            , sum_n_taxa_boo = FALSE
                            , sum_n_taxa_col = NULL
                            , sum_n_taxa_group_by = NULL
-                           , clean = FALSE
-                           , match_caps = FALSE) {
+                           , trim_ws = FALSE) {
 
   # DEBUG ----
   boo_DEBUG_tt <- FALSE
@@ -196,6 +191,8 @@ taxa_translate <- function(df_user = NULL
     sum_n_taxa_boo <- TRUE
     sum_n_taxa_col <- "N_TAXA"
     sum_n_taxa_group_by <- c("INDEX_NAME", "INDEX_CLASS", "SampleID", "TaxaID")
+    clean <- TRUE
+    match_caps <- TRUE
 
         ## OLD ----
     # # pick files
@@ -310,18 +307,11 @@ taxa_translate <- function(df_user = NULL
 
 
 
-  # Match_caps, Clean ----
+  # Munge1, trim_ws ----
   # 20240430, v1.0.2.9017, partial
-  # 20240528, v1.0.2.9025
-  if (match_caps & clean) {
-    # Munge, CAPS
-    df_official[, taxaid_official_match] <- toupper(df_official[, taxaid_official_match])
-    df_user[, taxaid_user] <- toupper(df_user[, taxaid_user])
-  }## IF ~ match_caps
-
-  if (clean) {
+  # 20240528, v1.0.2.9025 and 9026
+  if (trim_ws) {
     # Munge, clean
-    df_user[, taxaid_user] <- trimws(df_user[, taxaid_user])
     df_user[, taxaid_user] <- trimws(df_user[, taxaid_user]
                                      , whitespace = "[\\h\\v]")
   }## IF ~ clean
@@ -335,13 +325,6 @@ taxa_translate <- function(df_user = NULL
                     , sort = FALSE)
 
 
-
-
-
-
-
-
-
   if (boo_DEBUG_tt == TRUE) {
     testthat::expect_equal(nrow(df_user), nrow(df_merge))
   } ## IF ~ boo_DEBUG_tt
@@ -349,7 +332,7 @@ taxa_translate <- function(df_user = NULL
   # user taxa id will be gone after the merge
 
 
-  # Munge ----
+  # Munge2 ----
 
   ## new Col, match merge main ID to df_official----
   df_merge[, "Match_Official"] <- df_merge[, taxaid_official_match] %in%